+++ packages/subgraph/src/packages/addresslist/addresslist-voting.ts
@@ -2,17 +2,17 @@ import {
   Action,
   AddresslistVotingPlugin,
   AddresslistVotingProposal,
-  AddresslistVotingVoter,
   AddresslistVotingVote,
+  AddresslistVotingVoter,
 } from '../../../generated/schema';
 import {
-  VoteCast,
-  ProposalCreated,
-  ProposalExecuted,
-  VotingSettingsUpdated,
+  AddresslistVoting,
   MembersAdded,
   MembersRemoved,
-  AddresslistVoting,
+  ProposalCreated,
+  ProposalExecuted,
+  VoteCast,
+  VotingSettingsUpdated,
 } from '../../../generated/templates/AddresslistVoting/AddresslistVoting';
 import {RATIO_BASE, VOTER_OPTIONS, VOTING_MODES} from '../../utils/constants';
 import {getProposalId} from '../../utils/proposals';
@@ -44,7 +44,7 @@ export function _handleProposalCreated(
   proposalEntity.createdAt = event.block.timestamp;
   proposalEntity.creationBlockNumber = event.block.number;
   proposalEntity.allowFailureMap = event.params.allowFailureMap;
-  proposalEntity.potentiallyExecutable = false;
+  proposalEntity.approvalReached = false;
 
   let contract = AddresslistVoting.bind(pluginAddress);
   let proposal = contract.try_getProposal(pluginProposalId);
@@ -88,7 +88,7 @@ export function _handleProposalCreated(
       actionEntity.proposal = proposalId;
       actionEntity.save();
     }
-
+    proposalEntity.isSignaling = actions.length == 0;
     // totalVotingPower
     proposalEntity.totalVotingPower = contract.try_totalVotingPower(
       parameters.snapshotBlock
@@ -185,7 +185,7 @@ export function handleVoteCast(event: VoteCast): void {
         let minParticipationReached = castedVotingPower.ge(minVotingPower);
 
         // Used when proposal has ended.
-        proposalEntity.potentiallyExecutable =
+        proposalEntity.approvalReached =
           supportThresholdReached && minParticipationReached;
 
         // Used when proposal has not ended.
@@ -206,7 +206,7 @@ export function handleProposalExecuted(event: ProposalExecuted): void {
   let proposalEntity = AddresslistVotingProposal.load(proposalId);
   if (proposalEntity) {
     proposalEntity.executed = true;
-    proposalEntity.potentiallyExecutable = false;
+    proposalEntity.approvalReached = true;
     proposalEntity.executionDate = event.block.timestamp;
     proposalEntity.executionBlockNumber = event.block.number;
     proposalEntity.executionTxHash = event.transaction.hash;